added backend files as an interface simulating a backend in regards to #2
